# TechShopApp/TechShopApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "RestApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Dostawcy": {
      "get": {
        "tags": [
          "Dostawcy"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DostawcyForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DostawcyForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DostawcyForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dostawcy"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Dostawcy/{id}": {
      "get": {
        "tags": [
          "Dostawcy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DostawcyForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dostawcy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DostawcyForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Dostawcy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Produkty/{id}": {
      "get": {
        "tags": [
          "Produkty"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produkty"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Produkty"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ZamowioneProdukty": {
      "get": {
        "tags": [
          "ZamowioneProdukty"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZamowioneProduktyForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZamowioneProduktyForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZamowioneProduktyForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ZamowioneProdukty"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/ZamowioneProdukty/{zamowienieId}/{produktId}": {
      "get": {
        "tags": [
          "ZamowioneProdukty"
        ],
        "parameters": [
          {
            "name": "zamowienieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "produktId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZamowioneProduktyForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ZamowioneProdukty"
        ],
        "parameters": [
          {
            "name": "zamowienieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "produktId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ZamowioneProduktyForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ZamowioneProdukty"
        ],
        "parameters": [
          {
            "name": "zamowienieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "produktId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DostawcyForView": {
        "type": "object",
        "properties": {
          "DostawcaId": {
            "type": "integer",
            "format": "int32"
          },
          "Nazwa": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "Telefon": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "Email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "Adres": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "Email": {
            "type": "string",
            "nullable": true
          },
          "Haslo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProduktForView": {
        "type": "object",
        "properties": {
          "ProduktId": {
            "type": "integer",
            "format": "int32"
          },
          "Nazwa": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "Opis": {
            "type": "string",
            "nullable": true
          },
          "Cena": {
            "type": "number",
            "format": "double"
          },
          "StanMagazynowy": {
            "type": "integer",
            "format": "int32"
          },
          "KategoriaId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "KategoriaData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "Email": {
            "type": "string",
            "nullable": true
          },
          "Haslo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ZamowioneProduktyForView": {
        "type": "object",
        "properties": {
          "ZamowienieId": {
            "type": "integer",
            "format": "int32"
          },
          "ZamowienieData": {
            "type": "string",
            "nullable": true
          },
          "ProduktId": {
            "type": "integer",
            "format": "int32"
          },
          "ProduktData": {
            "type": "string",
            "nullable": true
          },
          "Ilosc": {
            "type": "integer",
            "format": "int32"
          },
          "CenaJednostkowa": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}